requirements.txt
fastapi
uvicorn[standard]
openai
chromadb
python-dotenv
